refactor(md): organized files in folders
