+++ app/scripts/modules/core/src/managed/artifactDetail/ArtifactDetail.less
