+++ app/scripts/modules/core/src/managed/artifactDetail/ArtifactDetailHeader.less
