+++ app/scripts/modules/core/src/managed/artifactDetail/ArtifactDetailHeader.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { Icon } from '../presentation';
-import { IManagedArtifactVersion } from '../domain';
+import { Icon } from '../../presentation';
+import { IManagedArtifactVersion } from '../../domain';
 
-import { getArtifactVersionDisplayName } from './displayNames';
+import { getArtifactVersionDisplayName } from '../displayNames';
 
 import './ArtifactDetailHeader.less';
