+++ app/scripts/modules/core/src/managed/artifactDetail/PreDeploymentRow.tsx
