+++ app/scripts/modules/core/src/managed/artifactDetail/constraints/ConstraintCard.less
