+++ app/scripts/modules/core/src/managed/artifactsList/ArtifactRow.less
