+++ app/scripts/modules/core/src/managed/artifactsList/ArtifactsList.tsx
@@ -7,14 +7,14 @@ import {
   IManagedArtifactVersion,
   IStatefulConstraint,
   StatefulConstraintStatus,
-} from '../domain/IManagedEntity';
-import { Icon, IconNames } from '../presentation';
+} from '../../domain/IManagedEntity';
+import { Icon, IconNames } from '../../presentation';
 
-import { isConstraintSupported, getConstraintIcon } from './constraints/constraintRegistry';
+import { isConstraintSupported, getConstraintIcon } from '../artifactDetail/constraints/constraintRegistry';
 
-import { ISelectedArtifactVersion } from './Environments';
-import { Pill } from './Pill';
-import { RelativeTimestamp } from './RelativeTimestamp';
+import { ISelectedArtifactVersion } from '../Environments';
+import { Pill } from '../Pill';
+import { RelativeTimestamp } from '../RelativeTimestamp';
 import { IStatusBubbleStackProps, StatusBubbleStack } from './StatusBubbleStack';
 
 import './ArtifactRow.less';
