+++ app/scripts/modules/core/src/managed/artifactsList/StatusBubbleStack.less
