+++ app/scripts/modules/core/src/managed/artifactsList/StatusBubbleStack.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { take } from 'lodash';
 
-import { IStatusBubbleProps, StatusBubble } from './StatusBubble';
+import { IStatusBubbleProps, StatusBubble } from '../StatusBubble';
 
 import './StatusBubbleStack.less';
