+++ app/scripts/modules/core/src/managed/environment/EnvironmentRow.less
