+++ app/scripts/modules/core/src/managed/resourceHistory/ManagedResourceDiffTable.tsx
