+++ app/scripts/modules/core/src/managed/resourceHistory/ManagedResourceHistoryModal.less
