+++ app/scripts/modules/core/src/managed/resourceHistory/ManagedResourceHistoryModal.tsx
@@ -14,12 +14,12 @@ import {
   usePollingData,
   usePrevious,
   showModal,
-} from '../presentation';
+} from '../../presentation';
 
 import { relativeTime, timestamp } from 'core/utils';
 import { IManagedResourceSummary, IManagedResourceDiff, IManagedResourceEvent } from 'core/domain';
 import { AccountTag } from 'core/account';
-import { ManagedReader } from './ManagedReader';
+import { ManagedReader } from '../ManagedReader';
 import { Spinner } from 'core/widgets';
 
 import { ManagedResourceDiffTable } from './ManagedResourceDiffTable';
